requirements.txt
fastapi
uvicorn
faiss-cpu
python-dotenv
openai
faiss-cpu
